movement started

+++ UDP_Client/UDP_Client/ImageObject.h
@@ -11,7 +11,7 @@ public:
 		physics->AddCollider(new AABB(sourceOffset, sourceSize * 0.2f));
 		renderer = new ImageRenderer(transform, texturepath, sourceOffset, sourceSize);
 	}
-	ImageObject(std::string texturepath, Vector2 sourceOffset, Vector2 sourceSize, int numColumns, int numRows, float frameTime, bool hasToLoop)
+	ImageObject(std::string texturepath, Vector2 sourceOffset, Vector2 sourceSize, unsigned short numColumns, unsigned short numRows, float frameTime, bool hasToLoop)
 		: Object() {
 		transform->size = sourceSize;
 		physics->AddCollider(new AABB(sourceOffset, sourceSize * 0.2f));
